# Conflicts:
#	CUE4Parse/UE4/Objects/Chaos/GeometryCollection/FGeometryCollectionNaniteData.cs
#	CUE4Parse/UE4/Objects/Chaos/GeometryCollection/FKeyType.cs
#	CUE4Parse/UE4/Objects/Chaos/GeometryCollection/FManagedArrayCollection.cs
#	CUE4Parse/UE4/Objects/Chaos/GeometryCollection/FValueType.cs
#	CUE4Parse/UE4/Objects/Chaos/GeometryCollection/TManangedArray.cs
#	CUE4Parse/UE4/Objects/Chaos/GeometryCollection/UGeometryCollection.cs

+++ CUE4Parse/UE4/Assets/Exports/Texture/UTextureCube.cs
@@ -28,7 +28,7 @@ public class UTextureCube : UTexture
         {
             var SizeX = Ar.Read<int>();
             var SizeY = Ar.Read<int>();
-            Format = (EPixelFormat) Ar.Read<int>();
+            Format = (EPixelFormat) Ar.Read<byte>();
             var numMips = Ar.Read<int>();
         }
